semaphore bug workarounds

+++ src/semaphore.sh
@@ -1,34 +1,31 @@
 #/bin/sh
 case $1 in
     setup)
-        mkdir -pv ${SEMAPHORE_CACHE_DIR}/apt/partial
+        export BUILDDIR=${HOME}/build/${BRANCH_NAME}
+        mkdir -pv ${SEMAPHORE_CACHE_DIR}/apt/archives/partial
         sudo cp -ruv /var/cache/apt ${SEMAPHORE_CACHE_DIR}/apt
         sudo rm -rfv /var/cache/apt
         sudo ln -sv ${SEMAPHORE_CACHE_DIR}/apt /var/cache/apt
         ;;
-    build-1)
+    build)
         sudo apt-get update
         sudo apt-get -fy install build-essential zlib1g-dev libsdl-mixer1.2-dev libsdl-image1.2-dev
-        make PLATFORM=linux64 PLATFORM_BIN=amd64 INSTDIR=${SEMAPHORE_CACHE_DIR}/${BRANCH_NAME}/linux/bin/amd64 CFLAGS=-m64 CXXFLAGS=-m64 LDFLAGS=-m64 -C src clean install || exit 1
+        make PLATFORM=linux64 PLATFORM_BIN=amd64 INSTDIR=linux/bin/amd64 CFLAGS=-m64 CXXFLAGS=-m64 LDFLAGS=-m64 -C src clean install || exit 1
         if [ "${BRANCH_NAME}" = "master" ]; then
             sudo apt-get -fy install binutils-mingw-w64 g++-mingw-w64
-            make PLATFORM=crossmingw64 PLATFORM_BIN=amd64 INSTDIR=${SEMAPHORE_CACHE_DIR}/${BRANCH_NAME}/windows/bin/amd64 CFLAGS=-m64 CXXFLAGS=-m64 LDFLAGS=-m64 -C src clean install || exit 1
-            make PLATFORM=crossmingw32 PLATFORM_BIN=x86 INSTDIR=${SEMAPHORE_CACHE_DIR}/${BRANCH_NAME}/windows/bin/x86 CFLAGS=-m32 CXXFLAGS=-m32 LDFLAGS=-m32 -C src clean install || exit 1
-        fi
-        ;;
-    build-2)
-        if [ "${BRANCH_NAME}" = "master" ]; then
+            make PLATFORM=crossmingw64 PLATFORM_BIN=amd64 INSTDIR=${BUILDDIR}/windows/bin/amd64 CFLAGS=-m64 CXXFLAGS=-m64 LDFLAGS=-m64 -C src clean install || exit 1
+            make PLATFORM=crossmingw32 PLATFORM_BIN=x86 INSTDIR=${BUILDDIR}/windows/bin/x86 CFLAGS=-m32 CXXFLAGS=-m32 LDFLAGS=-m32 -C src clean install || exit 1
             sudo dpkg --add-architecture i386
             sudo apt-get update
             sudo apt-get -fy remove zlib1g-dev libsdl1.2-dev libsdl-mixer1.2-dev libsdl-image1.2-dev libpng-dev
             sudo apt-get -fy autoremove
             sudo apt-get -fy install build-essential multiarch-support g++-multilib zlib1g-dev:i386 libsdl1.2-dev:i386 libsdl-mixer1.2-dev:i386 libsdl-image1.2-dev:i386 libpng-dev:i386
-            make PLATFORM=linux32 PLATFORM_BIN=x86 INSTDIR=${SEMAPHORE_CACHE_DIR}/${BRANCH_NAME}/linux/bin/x86 CFLAGS=-m32 CXXFLAGS=-m32 LDFLAGS=-m32 -C src clean install || exit 1
+            make PLATFORM=linux32 PLATFORM_BIN=x86 INSTDIR=${BUILDDIR}/linux/bin/x86 CFLAGS=-m32 CXXFLAGS=-m32 LDFLAGS=-m32 -C src clean install || exit 1
         fi
         ;;
     deploy)
-        if [ "${BRANCH_NAME}" = "master" ]; then
-            pushd ${SEMAPHORE_CACHE_DIR}/${BRANCH_NAME} || exit 1
+        if [ "${BRANCH_NAME}" = "master" ] && [ "${SEMAPHORE_THREAD_RESULT}" = "passed" ]; then
+            pushd ${BUILDDIR} || exit 1
             rm -fv version.txt windows.zip linux.tar.bz2
             pushd windows || exit 1
             zip -r ../windows.zip .
